Add Allure Support (Issue #6):
Add Allure dependency + Allure rest assured dependency
Add annotations for tests and for steps (methods inside controller).
Add custom assert methods
Add filter inside controller to save request/response to Allure attachments.
Remove prettyPrint() from code since now you can get data from Allure.

+++ src/main/java/models/Category.java
@@ -3,14 +3,14 @@ package models;
 import java.util.Objects;
 
 public class Category {
-    private Integer id;
+    private long id;
     private String name;
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -22,7 +22,7 @@ public class Category {
         this.name = name;
     }
 
-    public Category(Integer id, String name) {
+    public Category(long id, String name) {
         this.id = id;
         this.name = name;
     }
